IFS - [Refactor] Modifications thinking about cleanCode and good practices

+++ Vanguardium/Vanguardium.ApplicationService/Dtos/RabbitDtos/TransferProducerDto.cs
@@ -6,7 +6,7 @@ public sealed record TransferProducerDto
 {
     public Guid SenderId { get; init; }
     public Guid RecipientId { get; init; }
-    public int ValueForTransfer { get; init; }
+    public decimal ValueForTransfer { get; init; }
     public DateTime CreateDate { get; init; }
     public StatusTransfer StatusTransfer { get; init; }
 }
